Internal API

PowerNetworkMatrices.PowerNetworkMatrix — Type

Type PowerNetworkMatrix gathers all the different types of Matrices considered in this package

PowerNetworkMatrices.PowerNetworkMatrixKey — Type

Structure to store the keys of a power network matrix

Arguments

I<:Tuple: turple containing the indices of the matrix

PowerNetworkMatrices.PowerNetworkMatrixKeys — Type

Structure to store the keys of a power network matrix

Arguments

product_iter::Base.Iterators.ProductIterator{T} where T <: Tuple: iterator of the indices of the network power matrix

PowerNetworkMatrices.RowCache — Type

Structure used for saving the rows of the Virtual PTDF matrix.

Arguments

temp_cache::Dict{Int, Vector{Float64}}: Dictionay saving the row of the PTDF matrix

persistent_cache_keys::Set{Int}: Set listing the rows saved in temp_cache

max_cache_size::Int Defines the maximum allowed cache size (rows*row_size)

max_num_keys::Int Defines the maximum number of keys saved (rows of the matrix)

Base.eachindex — Method

eachindex(vptdf)

Gives the cartesian indexes of the PTDF matrix (same as the BA one).

Base.empty! — Method

empty!(cache)

Erases the cache.

Base.getindex — Method

getindex(cache, key)

Gets the row of the stored matrix in cache.

Arguments

cache::RowCache: cache where the row vector is going to be saved

key::Int: row number (corresponding to the enumerated branch index) related to the row vector.

Base.getindex — Method

getindex(vptdf, row, column)

Gets the value of the element of the PTDF matrix given the row and column indices corresponding to the branch and buses one respectively. If column is a Colon then the entire row is returned.

Arguments

vptdf::VirtualPTDF: VirtualPTDF struct where to evaluate and store the values.

row: Branch index.

column: Bus index. If Colon then get the values of the whole row.

Base.haskey — Method

haskey(cache, key)

Checks if key is present as a key of the dictionary in cache

Arguments

cache::RowCache: cache where data is stored.

key::Int: row number (corresponds to the enumerated branch index).

Base.isempty — Method

isempty(cache)

Check if cache is empty.

Base.isempty — Method

isempty(vptdf)

Checks if the any of the fields of VirtualPTDF is empty.

Base.length — Method

length(cache)

Shows the number of rows stored in cache

Base.setindex! — Method

setindex!(cache, val, key)

Allocates vector as row of the matrix saved in cache.

Arguments

cache::RowCache: cache where the row vector is going to be saved

val::Vector{Float64}: vector to be saved

key::Int: row number (corresponding to the enumerated branch index) related to the input row vector

Base.setindex! — Method

setindex!(_, _, _)

!!! STILL TO IMPLEMENT !!!

Base.setindex! — Method

setindex!(_, _, idx)

!!! STILL TO IMPLEMENT !!!

Base.size — Method

size(vptdf)

Gives the size of the whole PTDF matrix, not the number of rows stored.

PowerNetworkMatrices._binfo_check — Method

_binfo_check(binfo)

!!! MISSING DOCUMENTATION !!!

PowerNetworkMatrices._calculate_PTDF_matrix_DENSE — Method

_calculate_PTDF_matrix_DENSE(
    A,
    BA,
    ref_bus_positions,
    dist_slack
)

Funciton for internal use only.

Computes the PTDF matrix by means of the LAPACK and BLAS functions for dense matrices.

Arguments

A::Matrix{Int8}: Incidence Matrix

BA::Matrix{T} where {T <: Union{Float32, Float64}}: BA matrix

ref_bus_positions::Set{Int}: vector containing the indexes of the reference slack buses.

dist_slack::Vector{Float64}): vector containing the weights for the distributed slacks.

PowerNetworkMatrices._calculate_PTDF_matrix_KLU — Method

_calculate_PTDF_matrix_KLU(
    A,
    BA,
    ref_bus_positions,
    dist_slack
)

Funciton for internal use only.

Computes the PTDF matrix by means of the KLU.LU factorization for sparse matrices.

Arguments

A::SparseArrays.SparseMatrixCSC{Int8, Int}: Incidence Matrix

BA::SparseArrays.SparseMatrixCSC{Float64, Int}: BA matrix

ref_bus_positions::Set{Int}: vector containing the indexes of the reference slack buses.

dist_slack::Vector{Float64}: vector containing the weights for the distributed slacks.

PowerNetworkMatrices._calculate_PTDF_matrix_MKLPardiso — Method

_calculate_PTDF_matrix_MKLPardiso(
    A,
    BA,
    ref_bus_positions,
    dist_slack
)

Funciton for internal use only.

Computes the PTDF matrix by means of the MKL Pardiso for dense matrices.

Arguments

A::SparseArrays.SparseMatrixCSC{Int8, Int}: Incidence Matrix

BA::SparseArrays.SparseMatrixCSC{Float64, Int}: BA matrix

ref_bus_positions::Set{Int}: vector containing the indexes of the referece slack buses.

dist_slack::Vector{Float64}: vector containing the weights for the distributed slacks.

PowerNetworkMatrices.assing_reference_buses — Method

assing_reference_buses(subnetworks, ref_bus_positions)

!!! MISSING DOCUMENTATION !!!

PowerNetworkMatrices.calculate_ABA_matrix — Method

calculate_ABA_matrix(A, BA, ref_bus_positions)

Evaluates the ABA matrix given the System's Incidence matrix (A), BA matrix and reference bus positions.

Arguments

A::SparseArrays.SparseMatrixCSC{Int8, Int}: Incidence matrix.

BA::SparseArrays.SparseMatrixCSC{Float64, Int} BA matrix.

NOTE:

evaluates A with "calculateAmatrix", or extract A.data (if A::IncidenceMatrix)

evaluates BA with "calculateBAmatrix", or extract BA.data (if A::BA_Matrix)

PowerNetworkMatrices.calculate_A_matrix — Method

calculate_A_matrix(branches, buses)

Evaluates the Incidence matrix A given the branches and node of a System.

Arguments

branches: vector containing the branches of the considered system (should be AC branches).

buses::Vector{PSY.Bus}: vector containing the buses of the considered system.

NOTE:

the matrix features all the columns, including the ones related to the reference buses (each column is related to a system's bus).

PowerNetworkMatrices.calculate_BA_matrix — Method

calculate_BA_matrix(branches, bus_lookup)

Evaluates the transposed BA matrix given the System's banches, reference bus positions and bus_lookup.

Arguments

branches: vector containing the branches of the considered system (should be AC branches).

ref_bus_positions::Set{Int}: Vector containing the indexes of the columns of the BA matrix corresponding to the refence buses

bus_lookup::Dict{Int, Int}: dictionary mapping the bus numbers with their enumerated indexes.

PowerNetworkMatrices.calculate_PTDF_matrix_DENSE — Method

calculate_PTDF_matrix_DENSE(
    branches,
    nodes,
    bus_lookup,
    ref_bus_positions,
    dist_slack
)

Computes the PTDF matrix by means of the LAPACK and BLAS functions for dense matrices.

Arguments

branches: vector of the System AC branches

nodes::Vector{PSY.Bus}: vector of the System buses

bus_lookup::Dict{Int, Int}: dictionary mapping the bus numbers with their enumerated indexes.

dist_slack::Vector{Float64}: vector containing the weights for the distributed slacks.

PowerNetworkMatrices.calculate_PTDF_matrix_KLU — Method

calculate_PTDF_matrix_KLU(
    branches,
    nodes,
    bus_lookup,
    ref_bus_positions,
    dist_slack
)

Computes the PTDF matrix by means of the KLU.LU factorization for sparse matrices.

Arguments

branches: vector of the System AC branches

nodes::Vector{PSY.Bus}: vector of the System buses

bus_lookup::Dict{Int, Int}: dictionary mapping the bus numbers with their enumerated indexes.

dist_slack::Vector{Float64}: vector containing the weights for the distributed slacks.

PowerNetworkMatrices.calculate_PTDF_matrix_MKLPardiso — Method

calculate_PTDF_matrix_MKLPardiso(
    branches,
    nodes,
    bus_lookup,
    ref_bus_positions,
    dist_slack
)

Computes the PTDF matrix by means of the MKL Pardiso for dense matrices.

Arguments

branches: vector of the System AC branches

nodes::Vector{PSY.Bus}: vector of the System buses

bus_lookup::Dict{Int, Int}: dictionary mapping the bus numbers with their enumerated indexes.

dist_slack::Vector{Float64}: vector containing the weights for the distributed slacks.

PowerNetworkMatrices.calculate_adjacency — Method

calculate_adjacency(branches, buses, bus_lookup)

Evaluates the Adjacency matrix given the System's banches, buses and bus_lookup.

NOTE:

bus_lookup is a dictionary mapping the bus numbers (as shown in the Systems) with their enumerated indxes.

PowerNetworkMatrices.calculate_adjacency — Method

calculate_adjacency(branches, buses)

Evaluates the Adjacency matrix given the banches and buses of a given System.

Arguments

branches: vector containing the branches of the considered system (should be AC branches).

buses::Vector{PSY.Bus}: vector containing the buses of the considered system.

PowerNetworkMatrices.check_cache_size! — Method

check_cache_size!(cache)

Check saved rows in cache and deletes those ones that are not present in "persistentcachekeys".

PowerNetworkMatrices.find_slack_positions — Method

find_slack_positions(nodes)

Gets the indices of the reference (slack) buses. NOTE:

the indices corresponds to the columns of zeros belonging to the PTDF matrix.

BA and ABA matrix miss the columns related to the reference buses.

PowerNetworkMatrices.get_ac_branches — Method

get_ac_branches(sys)

Gets the AC branches from a given Systems.

PowerNetworkMatrices.get_bus_indices — Method

get_bus_indices(branch, bus_lookup)

Evaluates the bus indices for the given branch.

Arguments

branch: system's branch

bus_lookup: dictionary mapping the system's buses and branches

PowerNetworkMatrices.get_buses — Method

get_buses(sys)

Gets the buses from a given System

PowerNetworkMatrices.get_data — Method

get_data(mat)

returns the raw array data of the PowerNetworkMatrix

PowerNetworkMatrices.get_lookup — Method

get_lookup(mat)

returns the lookup tuple of the `PowerNetworkMatrix`. The first entry corresponds
to the first dimension and the second entry corresponds to the second dimension. For
instance in Ybus the first dimension is buses and second dimension is buses too, and in
PTDF the first dimension is branches and the second dimension is buses

PowerNetworkMatrices.get_tol — Method

get_tol(vptdf)

Gets the tolerance used for sparsifying the rows of the PTDF matrix

PowerNetworkMatrices.lookup_index — Method

lookup_index(i, lookup)

Gets bus indices to a certain branch index

PowerNetworkMatrices.lookup_index — Method

lookup_index(i, lookup)

Gets bus indices to a certain branch name

Arguments

i::PSY.ACBranch: Power System AC branch

lookup::Dict: Dictionary mapping branch and buses

PowerNetworkMatrices.lookup_index — Method

lookup_index(i, lookup)

Gets bus indices to a certain branch name

Arguments

i::PSY.ACBranch: Power System AC branch

lookup::Dict: Dictionary mapping branches and buses

PowerNetworkMatrices.make_ax_ref — Method

make_ax_ref(buses)

Evaluates the map linking the system's buses and branches.

Arguments

buses::AbstractVector{PSY.Bus}: system's buses

PowerNetworkMatrices.make_ax_ref — Method

make_ax_ref(ax)

Checkes if repetitions are present in the dictionary mapping buses and branches.

Arguments

ax::AbstractVector: generic abstract vector

PowerNetworkMatrices.make_entries_zero! — Method

make_entries_zero!(dense_array, tol)

Sets to zero every element of a Dense matrix if absolute values is below a certain tolerance.

Arguments

dense_array::Matrix{Float64}: input dense matrix.

tol::Float64: tolerance.

PowerNetworkMatrices.make_entries_zero! — Method

make_entries_zero!(sparse_array, tol)

Sets to zero every element of a Sparse matrix if absolute values is below a certain tolerance.

Arguments

sparse_array::SparseArrays.SparseMatrixCSC{Float64, Int}: input sparse array.

tol::Float64: tolerance for removing entries in the PTDF matrix.

PowerNetworkMatrices.make_entries_zero! — Method

make_entries_zero!(vector, tol)

Sets to zero every element of a Dense vector if absolute values is below a certain tolerance.

Arguments

vector::Vector{Float64}:input dense vector.

tol::Float64: tolerance.

PowerNetworkMatrices.purge_one! — Method

purge_one!(cache)

Deletes the row of of key key from the stored matrix in cache.

PowerNetworkMatrices.sparsify — Method

sparsify(dense_array, tol)

Return a sparse matrix given a dense one by dropping element whose absolute value is above a certain tolerance.

Arguments

dense_array::Matrix{Float64}`: input matrix (e.g., PTDF matrix).

tol::Float64: tolerance.

PowerNetworkMatrices.to_index — Method

to_index(A, idx)

Given the indices, gets the values of the power network matrix considered

PowerNetworkMatrices.validate_linear_solver — Method

validate_linear_solver(linear_solver)

Validates if the selected linear solver is supported.