Public API Reference

PowerNetworkMatrices.ABA_Matrix — Type

Structure containing the ABA matrix and other relevant data.

Arguments

data::SparseArrays.SparseMatrixCSC{Float64, Int}: the ABA matrix coming from the product between the Incidence Matrix A and the Matrix BA

axes<:NTuple{2, Dict}: Tuple containing two vectors, the first one contains the names of each line of the network (each one related to a row of the Matrix in "data"), the second one contains the names of each bus of the network (each one related to a column of the Matrix in "data")

lookup<:NTuple{2, Dict}: Tuple containing 2 Dictionaries mapping the number of rows and columns with the names of branches and buses

ref_bus_positions::Vector{Int}: Vector containing the indexes of the columns of the BA matrix corresponding to the refence buses

PowerNetworkMatrices.ABA_Matrix — Method

Builds the ABA matrix from a System

PowerNetworkMatrices.AdjacencyMatrix — Type

Nodal incidence matrix (Adjacency) is an N x N matrix describing a power system with N buses. It represents the directed connectivity of the buses in a power system.

The AdjacencyMatrix Struct is indexed using the Bus Numbers, no need for them to be sequential

Arguments

data::SparseArrays.SparseMatrixCSC{Int8, Int}: stores the incidence matrix

axes<:NTuple{2, Dict}: Tuple containing two vectors, the first one contains the names of each line of the network (each one related to a row of the Matrix in "data"), the second one contains the names of each bus of the network (each one related to a column of the Matrix in "data")

lookup<:NTuple{2, Dict}: Tuple containing 2 Dictionaries mapping the number of rows and columns with the names of branches and buses

ref_bus_positions::Vector{Int}: Vector containing the indexes of the columns of the BA matrix corresponding to the refence buses

PowerNetworkMatrices.AdjacencyMatrix — Method

Builds a AdjacencyMatrix from a collection of buses and branches. The return is an N x N AdjacencyMatrix Array indexed with the bus numbers.

Arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.AdjacencyMatrix — Method

Builds a AdjacencyMatrix from the system. The return is an N x N AdjacencyMatrix Array indexed with the bus numbers.

Arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.BA_Matrix — Type

Structure containing the BA matrix and other relevant data.

Arguments

data::SparseArrays.SparseMatrixCSC{Float64, Int}: the BA matrix coming from the product between the Incidence Matrix A and the Matrix of Susceptance B

axes<:NTuple{2, Dict}: Tuple containing two vectors, the first one contains the names of each line of the network (each one related to a row of the Matrix in "data"), the second one contains the names of each bus of the network (each one related to a column of the Matrix in "data")

lookup<:NTuple{2, Dict}: Tuple containing 2 Dictionaries mapping the number of rows and columns with the names of branches and buses

ref_bus_positions::Vector{Int}: Vector containing the indexes of the columns of the BA matrix corresponding to the refence buses

PowerNetworkMatrices.BA_Matrix — Method

Build the BA matrix from a given System

PowerNetworkMatrices.IncidenceMatrix — Type

Incidence matrix: shows connection between buses, defining lines

Arguments

data::SparseArrays.SparseMatrixCSC{Int8, Int}: the actual Incidence matrix.

axes<:NTuple{2, Dict}: Tuple containing two vectors (the first one showing the branches names, the second showing the buses numbers).

lookup<:NTuple{2, Dict}: Tuple containing two dictionaries, the first mapping the branches and buses with their enumerated indexes.

ref_bus_positions::Vector{Int}: vector containing the indices of the reference slack buses.

PowerNetworkMatrices.IncidenceMatrix — Method

Builds the Incidence matrix by evaluating the actual matrix and other relevant values.

PowerNetworkMatrices.LODF — Type

Builds the LODF matrix from a group of branches and nodes. The return is a LOLDF array indexed with the branch name.

Arguments

branches: vector of the System AC branches

nodes::Vector{PSY.Bus}: vector of the System buses

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses.

PowerNetworkMatrices.LODF — Type

Builds the LODF matrix from a system. The return is a LOLDF array indexed with the branch name.

Arguments

sys::PSY.System: Power Systems system

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses.

PowerNetworkMatrices.LODF — Type

Line Outage Distribution Factors (LODFs) are a sensitivity measure of how a change in a line's flow affects the flows on other lines in the system.

Arguments

data<:AbstractArray{Float64, 2}: the actual Incidence matrix.

axes<:NTuple{2, Dict}: Tuple containing two vectors (the first one showing the branches names, the second showing the buses numbers).

lookup<:NTuple{2, Dict}: Tuple containing two dictionaries, the first mapping the branches and buses with their enumerated indexes.

PowerNetworkMatrices.PTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and Branch names.

Fields

data<:AbstractArray{Float64, 2}: the actual Incidence matrix.

axes<:NTuple{2, Dict}: Tuple containing two vectors (the first one showing the branches names, the second showing the buses numbers).

lookup<:NTuple{2, Dict}: Tuple containing two dictionaries, the first mapping the branches and buses with their enumerated indexes.

subnetworks::Dict{Int, Set{Int}}: dictionary containing the set of bus indexes defining the subnetworks of the system.

tol::Base.RefValue{Float64}: tolerance used for sportifying the matrix (dropping element whose absolute value is below this threshold).

PowerNetworkMatrices.PTDF — Method

Builds the PTDF matrix from a group of branches and nodes. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

branches: vector of the System AC branches

nodes::Vector{PSY.Bus}: vector of the System buses

dist_slack::Vector{Float64}: vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

linear_solver::String: Linear solver to be used. Options are "Dense", "KLU" and "MKLPardiso

tol::Float64: Tolerance to eliminate entries in the PTDF matrix (default eps())

PowerNetworkMatrices.PTDF — Method

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

A::IncidenceMatrix: Incidence Matrix (full structure)

BA::BA_Matrix: BA matrix (full structure)

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses.

linear_solver::String: Linear solver to be used. Options are "Dense", "KLU" and "MKLPardiso.

tol::Float64: Tolerance to eliminate entries in the PTDF matrix (default eps()).

PowerNetworkMatrices.PTDF — Method

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

sys::PSY.System: Power Systems system

PowerNetworkMatrices.VirtualPTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and branch names.

Arguments

K::KLU.KLUFactorization{Float64, Int}: LU factorization matrices of the ABA matrix, evaluated by means of KLU

BA::SparseArrays.SparseMatrixCSC{Float64, Int}: BA matric

ref_bus_positions::Vector{Int}: Vector containing the indexes of the columns of the BA matrix corresponding to the refence buses

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses.

axes<:NTuple{2, Dict}: Tuple containing two vectors (the first one showing the branches names, the second showing the buses numbers).

lookup<:NTuple{2, Dict}: Tuple containing two dictionaries, the first mapping the branches and buses with their enumerated indexes.

temp_data::Vector{Float64}: temporary vector for internal use.

cache::RowCache: cache were PTDF rows are stored.

subnetworks::Dict{Int, Set{Int}}: dictionary containing the subsets of buses defining the different subnetwork of the system.

tol::Base.RefValue{Float64}: tolerance related to scarification and values to drop.

PowerNetworkMatrices.VirtualPTDF — Method

Builds the PTDF matrix from a group of branches and buses. The return is a PTDF array indexed with the branch numbers.

Arguments

branches: Vector of the system's AC branches.

buses::Vector{PSY.Bus}: Vector of the system's buses.

dist_slack::Vector{Float64} = Float64[]: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses.

tol::Float64 = eps(): Tolerance related to sparsification and values to drop.

max_cache_size::Int: max cache size in MiB (inizialized as MAXCACHESIZE_MiB).

persistent_lines::Vector{String}: line to be evaluated as soon as the VirtualPTDF is created (initialized as empty vector of strings).

PowerNetworkMatrices.VirtualPTDF — Method

Builds the Virtual PTDF matrix from a system. The return is a VirtualPTDF struct with an empty cache.

PowerNetworkMatrices.Ybus — Type

Builds a Ybus from a collection of buses and branches. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.Ybus — Type

Nodal admittance matrix (Ybus) is an N x N matrix describing a power system with N buses. It represents the nodal admittance of the buses in a power system.

The Ybus Struct is indexed using the Bus Numbers, no need for them to be sequential

PowerNetworkMatrices.Ybus — Method

Builds a Ybus from the system. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network

PowerNetworkMatrices.find_subnetworks — Method

find_subnetworks(
    M::AdjacencyMatrix
) -> Dict{Int64, Set{Int64}}

Evaluates subnetworks by looking for the subsets of nodes connected each other, but not connected with nodes of other subsets.

PowerNetworkMatrices.find_subnetworks — Method

find_subnetworks(
    M::SparseArrays.SparseMatrixCSC,
    bus_numbers::Vector{Int64}
) -> Dict{Int64, Set{Int64}}

Finds the subnetworks present in the considered System. This is evaluated by taking a the ABA or Adjacency Matrix.

Keyword arguments

M::SparseArrays.SparseMatrixCSC: input sparse matrix.

bus_numbers::Vector{Int}: vector containing the indices of the system's buses.

PowerNetworkMatrices.validate_connectivity — Method

validate_connectivity(M::AdjacencyMatrix) -> Bool

Validates connectivity by checking that the number of subnetworks is 1 (fully connected network).

PowerNetworkMatrices.validate_connectivity — Method

validate_connectivity(sys::PowerSystems.System) -> Bool

Checks the network connectivity of the system.

Keyword arguments

connectivity_method::Function = goderya_connectivity: Specifies the method used as Goderya's algorithm (goderya_connectivity) or depth first search/network traversal (dfs_connectivity)

Note that the default Goderya method is more efficient, but is resource intensive and may not scale well on large networks.