Internal API

PowerNetworkMatrices.get_data — Method

get_data(
    mat::PowerNetworkMatrices.PowerNetworkMatrix
) -> Dict{Int64, Array{Float64, N} where N}

returns the raw array data of the PowerNetworkMatrix

PowerNetworkMatrices.get_lookup — Method

get_lookup(
    mat::PowerNetworkMatrices.PowerNetworkMatrix
) -> Tuple{Dict, Dict}

returns the lookup tuple of the `PowerNetworkMatrix`. The first entry corresponds
to the first dimension and the second entry corresponds to the second dimension. For
instance in Ybus the first dimension is buses and second dimension is buses too, and in
PTDF the first dimension is branches and the second dimension is buses