Public API Reference

PowerNetworkMatrices.AdjacencyMatrix — Type

Nodal incidence matrix (Adjacency) is an N x N matrix describing a power system with N buses. It represents the directed connectivity of the buses in a power system.

The AdjacencyMatrix Struct is indexed using the Bus Numbers, no need for them to be sequential

PowerNetworkMatrices.AdjacencyMatrix — Method

Builds a AdjacencyMatrix from a collection of buses and branches. The return is an N x N AdjacencyMatrix Array indexed with the bus numbers.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.AdjacencyMatrix — Method

Builds a AdjacencyMatrix from the system. The return is an N x N AdjacencyMatrix Array indexed with the bus numbers.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.LODF — Type

Line Outage Distribution Factors (LODFs) are a sensitivity measure of how a change in a line’s flow affects the flows on other lines in the system.

PowerNetworkMatrices.LODF — Type

Builds the LODF matrix from a group of branches and nodes. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerNetworkMatrices.LODF — Type

Builds the LODF matrix from a system. The return is a LOLDF array indexed with the branch name.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerNetworkMatrices.PTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and branch names

PowerNetworkMatrices.PTDF — Method

Builds the PTDF matrix from a group of branches and nodes. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

linear_solver::String: Linear solver to be used. Options are "Dense", "KLU" and "MKLPardiso

tol::Float64: Tolerance to eliminate entries in the PTDF matrix (default eps())

PowerNetworkMatrices.PTDF — Method

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

linear_solver::String: Linear solver to be used. Options are "Dense", "KLU" and "MKLPardiso

tol::Float64: Tolerance to eliminate entries in the PTDF matrix (default eps())

PowerNetworkMatrices.VirtualPTDF — Type

Power Transfer Distribution Factors (PTDF) indicate the incremental change in real power that occurs on transmission lines due to real power injections changes at the buses.

The PTDF struct is indexed using the Bus numbers and branch names

PowerNetworkMatrices.VirtualPTDF — Method

Builds the PTDF matrix from a group of branches and nodes. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerNetworkMatrices.VirtualPTDF — Method

Builds the PTDF matrix from a system. The return is a PTDF array indexed with the bus numbers.

Keyword arguments

dist_slack::Vector{Float64}: Vector of weights to be used as distributed slack bus. The distributed slack vector has to be the same length as the number of buses

PowerNetworkMatrices.Ybus — Type

Nodal admittance matrix (Ybus) is an N x N matrix describing a power system with N buses. It represents the nodal admittance of the buses in a power system.

The Ybus Struct is indexed using the Bus Numbers, no need for them to be sequential

PowerNetworkMatrices.Ybus — Type

Builds a Ybus from a collection of buses and branches. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network using Goderya's algorithm

PowerNetworkMatrices.Ybus — Method

Builds a Ybus from the system. The return is a Ybus Array indexed with the bus numbers and the branch names.

Keyword arguments

check_connectivity::Bool: Checks connectivity of the network

PowerNetworkMatrices.validate_connectivity — Method

validate_connectivity(sys::PowerSystems.System) -> Bool

Checks the network connectivity of the system.

Keyword arguments

connectivity_method::Function = goderya_connectivity: Specifies the method used as Goderya's algorithm (goderya_connectivity) or depth first search/network traversal (dfs_connectivity)

Note that the default Goderya method is more efficient, but is resource intensive and may not scale well on large networks.